spplan bug fix and code cleanup
